Remove getSystemE2ETestProvider from shared utils file, moved into standalone file
