+++ packages/web3/test/e2e/get_accounts.test.ts
@@ -15,11 +15,8 @@ You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 import Web3 from '../../src';
-import {
-	closeOpenConnection,
-	getSystemE2ETestProvider,
-	getSystemTestBackend,
-} from '../shared_fixtures/system_tests_utils';
+import { getSystemE2ETestProvider } from './get_system_e2e_test_provider';
+import { closeOpenConnection, getSystemTestBackend } from '../shared_fixtures/system_tests_utils';
 
 describe(`${getSystemTestBackend()} tests - getAccounts`, () => {
 	const provider = getSystemE2ETestProvider();
