+++ packages/web3/test/e2e/get_block.test.ts
@@ -20,11 +20,8 @@ import { blockSchema } from 'web3-eth';
 import { Transaction } from 'web3-types';
 
 import Web3 from '../../src';
-import {
-	closeOpenConnection,
-	getSystemE2ETestProvider,
-	getSystemTestBackend,
-} from '../shared_fixtures/system_tests_utils';
+import { getSystemE2ETestProvider } from './get_system_e2e_test_provider';
+import { closeOpenConnection, getSystemTestBackend } from '../shared_fixtures/system_tests_utils';
 import { toAllVariants } from '../shared_fixtures/utils';
 
 describe(`${getSystemTestBackend()} tests - getBlock`, () => {
