+++ scripts/system_tests_utils.ts
@@ -54,7 +54,6 @@ import { NonPayableMethodObject } from 'web3-eth-contract';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import HttpProvider from 'web3-providers-http';
 import accountsString from './accounts.json';
-import secrets from '../.secrets.json';
 
 /**
  * Get the env variable from Cypress if it exists or node process
@@ -88,16 +87,6 @@ export const isBrowser: boolean = ['chrome', 'firefox'].includes(getSystemTestEn
 export const getSystemTestMnemonic = (): string => getEnvVar('WEB3_SYSTEM_TEST_MNEMONIC') ?? '';
 
 export const getSystemTestBackend = (): string => getEnvVar('WEB3_SYSTEM_TEST_BACKEND') ?? '';
-export const getSystemE2ETestProvider = (): string => {
-	if (process.env.WEB3_SYTEM_TEST_MODE === 'http') {
-		return getSystemTestBackend() === 'sepolia'
-			? process.env.INFURA_SEPOLIA_HTTP ?? secrets.SEPOLIA.HTTP
-			: process.env.INFURA_MAINNET_HTTP ?? secrets.SEPOLIA.HTTP;
-	}
-	return getSystemTestBackend() === 'sepolia'
-		? process.env.INFURA_SEPOLIA_WS ?? secrets.SEPOLIA.WS
-		: process.env.INFURA_MAINNET_WS ?? secrets.MAINNET.WS;
-};
 
 export const createAccount = _createAccount;
